refactor(solver): Card 4b module 1 — carve out the native spatial kernel
`python/discopt/solver.py` becomes the package `python/discopt/solver/` (git
rename recorded, so `git log --follow` is preserved) and the first module is
carved out: `solver/native_kernel.py`, the eight-function #764
native-spatial-kernel cluster — engagement gate, feature safety, seed
construction and rigorous seed verification, and the one-attempt driver — plus
its three constants.

Provably a pure move: 7 of the 8 moved functions are AST-identical to their
pre-move selves, and the 8th (`_try_native_spatial_kernel`) differs by exactly
three added lines, a deferred `from discopt.solver import _unpack_solution` that
breaks the package cycle. All 124 stay-behind functions are AST-identical.
solver/__init__.py 18,211 -> 17,569 lines; native_kernel.py 796.

Why this cluster and not one of the five the card names: the entry experiment
(plan §6) measured the coupling. `setup`/`reformulate`/`root`/`spatial_loop`/
`results` are inline statement blocks of `solve_model`, not functions — they
share a closure of 200+ locals, so moving one is a signature-design problem, not
a move. Of the movable functions, every engine reaches 13-44 module-level names
(`_solve_nlp_bb`: 44, including `solve_model` itself); the native-kernel cluster
reaches exactly one.

The refactor hazard the card did not anticipate, and the rule adopted for it:
30+ `discopt.solver` attributes are monkeypatched by the test suite. A pure move
plus a re-export leaves `monkeypatch.setattr(discopt.solver, "_f", fake)`
succeeding and doing nothing, because the moved call site resolves `_f` in its
own module — and one such seam is a negative assertion (`must_not_run`) that
would then pass vacuously. So: a moved function that any test *patches* is NOT
re-exported (a stale patch raises AttributeError instead of silently no-op'ing);
one that tests only *call* may be re-exported with the explicit `x as x` form.
`_native_kernel_seed` and `_native_kernel_seed_candidates` are therefore not
re-exported, their three test sites now patch `discopt.solver.native_kernel`, and
`test_patched_seed_helpers_are_not_re_exported_from_discopt_solver` makes the
policy a standing test rather than a comment.

Regime N. `panel_baseline.py --check reports/panel_baseline_f154dcff.json` on
this tree, gated separately from Card 4a's: comparisons executed 255
(node_count 85, certified objective 85, status 85) over 85 comparable of 119
rows — PASS, no node-count or certified-objective drift. 2078.4 s, load start
1.08 peak 1.93; 17 non-comparable budget-dependent rows reported and not gating.
`pytest python/tests/test_spatial_native_kernel.py
python/tests/test_902_native_seed_general_integers.py` — 68 passed.

One post-panel edit is in this commit and is docstring-only: routing.py's header
said "roughly thirty gates spread over 2,900 lines"; the gates span ~4,060 lines,
re-derived and corrected (CLAUDE.md: numbers in a doc are read as measurements).
It cannot affect a node count.

Contributes to Phase 4 of docs/dev/consolidation-plan-2026-07-28.md; the plan's
Phase 4 status block and three §6 entries (Step 0 re-runs, Card 4a, the Card 4b
entry experiment) land with this commit.
